+++ api/src/reportcreator_api/tests/test_api.py
@@ -240,11 +240,11 @@ def guest_urls():
         ('projecttype create private', lambda s, c: c.post(reverse('projecttype-list'), data=c.get(reverse('projecttype-detail', kwargs={'pk': s.project_type.pk})).data | {'scope': ProjectTypeScope.PRIVATE})),
         ('projecttype import private', lambda s, c: c.post(reverse('projecttype-import'), data={'file': export_archive(s.project_type), 'scope': ProjectTypeScope.PRIVATE}, format='multipart')),
         *projecttype_viewset_urls(get_obj=lambda s: s.project_type, list=True, read=True),
-        *projecttype_viewset_urls(get_obj=lambda s: s.project_type_customized, read=True, write=True),
+        *projecttype_viewset_urls(get_obj=lambda s: s.project_type_customized, read=True, write=False),
         *projecttype_viewset_urls(get_obj=lambda s: s.project_type_snapshot, read=True),
         *projecttype_viewset_urls(get_obj=lambda s: s.project_type_private, read=True, write=True),
 
-        *project_viewset_urls(get_obj=lambda s: s.project, list=True, read=True, write=True, destory=False, update=False),
+        *project_viewset_urls(get_obj=lambda s: s.project, list=True, read=True, write=False, destory=False, update=False),
         *project_viewset_urls(get_obj=lambda s: s.project_readonly, read=True),
 
         *viewset_urls('archivedproject', get_kwargs=lambda s, detail: {'pk': s.archived_project.pk} if detail else {}, list=True, retrieve=True),
@@ -257,7 +257,8 @@ def regular_user_urls():
     return [
         *viewset_urls('pentestuser', get_kwargs=lambda s, detail: {'pk': s.user_other.pk} if detail else {}, retrieve=True),
 
-        *project_viewset_urls(get_obj=lambda s: s.project, create=True, update=True, destory=True),
+        *project_viewset_urls(get_obj=lambda s: s.project, create=True, update=True, write=True, destory=True),
+        *projecttype_viewset_urls(get_obj=lambda s: s.project_type_customized, write=True),
         ('pentestproject readonly', lambda s, c: c.put(reverse('pentestproject-readonly', kwargs={'pk': s.project.pk}), data={'readonly': True})),
         *project_viewset_urls(get_obj=lambda s: s.project_readonly, destory=True),
         ('pentestproject readonly', lambda s, c: c.put(reverse('pentestproject-readonly', kwargs={'pk': s.project_readonly.pk}), data={'readonly': False})),
@@ -452,6 +453,8 @@ def test_api_requests(username, name, perform_request, initialize_dependencies, 
             GUEST_USERS_CAN_CREATE_PROJECTS=False,
             GUEST_USERS_CAN_DELETE_PROJECTS=False,
             GUEST_USERS_CAN_UPDATE_PROJECT_SETTINGS=False,
+            GUEST_USERS_CAN_EDIT_PROJECTS=False,
+            GUEST_USERS_CAN_SEE_ALL_USERS=False,
             ARCHIVING_THRESHOLD=1,
             AUTHLIB_OAUTH_CLIENTS={
                 'dummy': {